Bugfix: Property is a float and can not be converted to a double

+++ java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/shared/RaceStatusDTO.java
@@ -5,7 +5,7 @@ import com.sap.sailing.domain.common.TrackedRaceState;
 
 public class RaceStatusDTO implements IsSerializable {
     public TrackedRaceState status;
-    public double loadingProgress;
+    public float loadingProgress;
     
     @Override
     public String toString() {
